Metaprogram methods into anonymous module to fix respond_to
Why
When I created a policy that subclassed from the base policy and wanted
to write tests to see if my new policy responded to what I thought it
should, `respond_to?` returned false even though the policy actually did
respond to the action in question. I created a `grade` action_on_object
and asked if it responded to `#grade?` and it returned false even though
it actually worked.

What
First possibility I saw was to make sure to define a
`responds_to_missing?` to go along with our `method_missing` as seen
[here](https://makandracards.com/makandra/9821-when-overriding-method_missing-remember-to-override-respond_to_missing-as-well).

But I was curious if we could just eliminate method_missing. The main
catch there is that we wanted a way that we could override the actions
in a base class without using the `alias_method` do-si-do and if we had
a method of the same name in the base class, then `define_method` would
not define the method in the super class because it was already
defined... and therefore we would get an error when trying to call
`super` in the override method.

To solve that, it sounds like we can use modules to place the methods
higher in the inheritance hierarchy than the subclass... allowing us to
define the methods straight in the module, override them in a subclass
while calling super, and they would also return true when calling
`respond_to?` because we would be able to ditch method_missing. Move
info on this technique in [this blog
post](https://thepugautomatic.com/2013/07/dsom/)